backend/pyproject.toml
[project]
name = "researchpulse"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "google-genai>=0.8.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "flower>=2.0.1",
    "chromadb>=0.5.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "sqlalchemy>=2.0.30",
    "aiosqlite>=0.20.0",
    "feedparser>=6.0.11",
    "tenacity>=8.3.0",
    "tiktoken>=0.7.0",
    "sentence-transformers>=3.0.0",
    "python-multipart>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
